# Improve performance using thread