--- a/lib/github/nippou/concerns/string_markdown.rb
+++ b/lib/github/nippou/concerns/string_markdown.rb
@@ -3,5 +3,10 @@ module StringMarkdown
     def markdown_escape
       self.gsub(/([`<>])/, '\\\\\1')
     end
+
+    def html_url_as_nippou
+      self =~ /^\* \[.+\]\((.+)\)/
+      $1
+    end
   end
 end