### Before

**19.019 sec** in the case of single thread

```
$ time bundle exec ruby bin/github-nippou -s 20160327
(17 items)
bundle exec ruby bin/github-nippou -s 20160327  1.52s user 0.26s system 9% cpu 19.019 total
```

### After

**7.006 sec** in the case of 5 threads (default)

```
$ time bundle exec ruby bin/github-nippou -s 20160327
(17 items)
bundle exec ruby bin/github-nippou -s 20160327  1.68s user 0.28s system 28% cpu 7.006 total
```

**5.525 sec** in the case of 10 threads

```
$ $ time GITHUB_NIPPOU_THREAD_NUM=10 bundle exec ruby bin/github-nippou -s 20160327
(17 items)
GITHUB_NIPPOU_THREAD_NUM=10 bundle exec ruby bin/github-nippou -s 20160327  1.55s user 0.27s system 32% cpu 5.525 total
```